pyproject.toml
[tool.poetry]
name = "tatarubot2-kook"
version = "0.1"
description = "FF14 bot Tataru for KOOK"
authors = ["aaron-li"]
readme = "docs/README.md"

[tool.poetry.dependencies]
python = "^3.7.3"
nonebot2 = "^2.0.0-beta.1"
nonebot-adapter-onebot = "*"
aiohttp = "*"
Pillow = "*"
icalendar = "*"
emoji = "*"
aiohttp_socks = "*"

[tool.poetry.dev-dependencies]
nb-cli = "^0.6.0"

[tool.nonebot]
plugins = []
plugin_dirs = ["tatarubot2"]

[build-system]
requires = ["poetry_core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
